Adding categories
Now user can create categories and assign different designs to them,
they will display as filtered in both retail and wholesale sections.
Played with layouts a bit, finished responsiveness and the bit where
page info is displayed in footer instead of hard-coded text.

There are some final bits to finish, but generally our theme is in
hand-in-ready state :smiley: